+++ lib/ndctl/libndctl.h
@@ -156,9 +156,9 @@ struct ndctl_cmd *ndctl_bus_cmd_new_ars_status(struct ndctl_cmd *ars_cap);
 unsigned int ndctl_cmd_ars_cap_get_size(struct ndctl_cmd *ars_cap);
 unsigned int ndctl_cmd_ars_in_progress(struct ndctl_cmd *ars_status);
 unsigned int ndctl_cmd_ars_num_records(struct ndctl_cmd *ars_stat);
-unsigned int ndctl_cmd_ars_get_record_addr(struct ndctl_cmd *ars_stat,
+unsigned long long ndctl_cmd_ars_get_record_addr(struct ndctl_cmd *ars_stat,
 		unsigned int rec_index);
-unsigned int ndctl_cmd_ars_get_record_len(struct ndctl_cmd *ars_stat,
+unsigned long long ndctl_cmd_ars_get_record_len(struct ndctl_cmd *ars_stat,
 		unsigned int rec_index);
 struct ndctl_cmd *ndctl_dimm_cmd_new_vendor_specific(struct ndctl_dimm *dimm,
 		unsigned int opcode, size_t input_size, size_t output_size);
